backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Dependencies for django-step-by-step"
authors = ["Brian Caffey <briancaffey2010@gmail.com>"]

[tool.poetry.dependencies]
python = "==3.9.13"

# Update to Django 3.2 requires additional changes
Django = "~3.2.0"
djangorestframework = "^3.13.1"

# redis
redis = "^4.1.0"

# celery
celery= "^5.2.3"

# for OpenAPI
uritemplate = "^3.0.1"
coreapi = "^2.3.3"
pyyaml = "^5.4.1"

# graphql
graphene-django = "^2.15.0"
django-graphql-jwt = "^0.3.4"
# TODO: watch this issue
# https://github.com/flavors/django-graphql-jwt/issues/241
PyJWT = ">2"
gunicorn = "^20.1.0"
# psycopg2-binary = { version = "^2.8.6", markers = "platform_machine == 'x86_64'" }
psycopg2-binary = "^2.9.3"
aws-secretsmanager-caching = "^1.1.1"
Pillow = "^9.0.0"
graphene-file-upload = "^1.3.0"
django-storages = "^1.12.3"
dj-database-url = "^0.5.0"
whitenoise = "^5.3.0"
boto3 = "^1.20.31"
django-cors-headers = "^3.10.0"
djangorestframework-simplejwt = "^5.0.0"

# sentry SDK
sentry-sdk = "^1.5.1"

# flower
flower = "^1.0"

[tool.poetry.dev-dependencies]

# test dependencies
black =  { version = "^20.8b1", markers = "platform_machine == 'x86_64'" }
flake8 = "^4.0.1"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pytest-django = "^4.5.2"

# factory
factory-boy = "^3.2.1"

# dev dependencies

django-debug-toolbar = "^3.2.4"

# for runserver_plus
django-extensions = "^3.1.5"
Werkzeug = "^2.0.2"

# for Jupyter notebooks
ipython = "^7.31.0"
# jupyter's dependency on argon2-cffi can't be installed on Apple M1
jupyter = { version = "^1.0.0", markers = "platform_machine == 'x86_64'" }
notebook = "^6.4.12"

# OpenAPI
pyyaml = "^5.3.1"
argh = "^0.26.2"

# faker
faker = "^11.3.0"

# django-silk
django-silk = "^4.2.0"

# pyparsing
# poetry 1.0.0 requires pyparsing<3.0,>=2.2 but packaging tries to install a 3.0.6 which is not compatible
pyparsing = "^2.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
